+++ NEWS.md
@@ -2,9 +2,11 @@
 
 #### New features
 * Add functions that creates HTML file base on `DistributionModel`.
+* Added new engine `engine_scampr` for model-based integration.
 
 #### Minor improvements and bug fixes
 * Small fixes to ensure `boruta` filtering works (again)?
+* Small fix to parameter in `train()` #102 @jeffreyhanson
 
 # ibis.iSDM 0.1.2